agentic-compliance-mapping/requirements.txt

# Core Dependencies
boto3==1.34.144
snowflake-connector-python==3.11.0
snowflake-sqlalchemy==1.6.0
langchain==0.2.11
langchain-community==0.2.10
langchain-openai==0.1.17
langgraph==0.1.9

# Document Processing
PyPDF2==3.0.1
pdf2image==1.17.0
pytesseract==0.3.10
unstructured==0.14.10
python-multipart==0.0.9

# ML and NLP
spacy==3.7.5
transformers==4.42.4
torch==2.3.1
sentence-transformers==3.0.1
numpy==1.26.4
pandas==2.2.2

# API and Web Framework
fastapi==0.111.0
uvicorn==0.30.1
pydantic==2.8.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Visualization and Reporting
plotly==5.22.0
matplotlib==3.9.1
seaborn==0.13.2
reportlab==4.2.2
openpyxl==3.1.5
jinja2==3.1.4

# Utilities
python-dotenv==1.0.1
requests==2.32.3
aiofiles==23.2.1
asyncio==3.4.3
uuid==1.30
python-json-logger==2.0.7
structlog==24.2.0

# Testing
pytest==8.2.2
pytest-asyncio==0.23.7
pytest-cov==5.0.0
moto==5.0.9
httpx==0.27.0

# Development
black==24.4.2
flake8==7.1.0
mypy==1.10.1
pre-commit==3.7.1
